server/requirements.txt
google-generativeai
Flask
flask-cors
Flask-Mail